VM: merge strings.hpp and tuples.hpp into layouts.hpp, because those header files were so tiny
