+++ vm/master.hpp
@@ -109,7 +109,6 @@ namespace factor { struct factor_vm; }
 #include "mark_bits.hpp"
 #include "free_list.hpp"
 #include "fixup.hpp"
-#include "tuples.hpp"
 #include "free_list_allocator.hpp"
 #include "write_barrier.hpp"
 #include "object_start_map.hpp"
@@ -118,7 +117,6 @@ namespace factor { struct factor_vm; }
 #include "data_heap.hpp"
 #include "code_heap.hpp"
 #include "gc.hpp"
-#include "strings.hpp"
 #include "float_bits.hpp"
 #include "io.hpp"
 #include "image.hpp"
